+++ go.mod
@@ -4,6 +4,7 @@ go 1.17
 
 require (
 	github.com/disintegration/imaging v1.6.2
+	github.com/essentialkaos/depsy v1.0.0
 	github.com/essentialkaos/ek/v12 v12.60.0
 )
